Count mails

+++ email_download/email_download.py
@@ -6,19 +6,10 @@ import sys
 
 from tqdm import tqdm
 from imap_tools import MailBox
-from imap_tools.message import EmailAddress
 import click
 import shutil
 from operator import attrgetter
-
-
-class HashableEmailAddress(EmailAddress):
-    def __init__(self, value: EmailAddress):
-        super().__init__(value.name, value.email)
-        self.domain = value.email.split('@')[-1]
-
-    def __hash__(self):
-        return hash(self.email)
+from hashable_mail import HashableEmailAddress
 
 
 @click.command()
@@ -27,7 +18,9 @@ class HashableEmailAddress(EmailAddress):
 @click.option('--password', prompt='Password', hide_input=True, help='The IMAP server password.')
 @click.option('--directory', default="saved_emails", prompt='Save Directory',
               help='Directory to save downloaded emails.')
-def main(host, user, password, directory):
+@click.option('--delete-contents', default=False, help='Delete the contents in the save directory..')
+@click.option('--overwrite', default=True, help='Overwrite the files in the save directory.')
+def main(host, user, password, directory, delete_contents, overwrite):
     """Download emails from the specified IMAP server and save them to the given directory."""
 
     def clear_directory(directory):
@@ -52,18 +45,20 @@ def main(host, user, password, directory):
         return filename
 
     if os.path.exists(directory) and os.listdir(directory):  # Checking if directory is not empty
-        if click.confirm(f"'{directory}' is not empty. Do you want to delete its contents?", default=False):
+        if delete_contents or click.confirm(f"'{directory}' is not empty. Do you want to delete its contents?",
+                                            default=False):
             clear_directory(directory)
     elif not os.path.exists(directory):
         os.makedirs(directory)
 
-    senders = set()
+    senders = dict()
 
     with MailBox(host).login(user, password) as mailbox:
         total_emails = len(mailbox.numbers())
         with tqdm(total=total_emails, dynamic_ncols=True) as pbar:
             for msg in mailbox.fetch():
-                senders.add(HashableEmailAddress(msg.from_values))
+                email_address = HashableEmailAddress(msg.from_values)
+                senders[email_address] = senders.get(email_address, 0) + 1
 
                 formatted_from = f"{msg.from_values.full[:30]:<30}"
                 formatted_subject = f"{msg.subject[:40]:<40}"
@@ -72,22 +67,26 @@ def main(host, user, password, directory):
 
                 filepath = os.path.join(directory, generate_filename(msg))
 
-                with open(filepath, 'wb') as file:
-                    try:
-                        email_object = msg.obj.as_string().encode('utf-8')
-                        file.write(email_object)
-                    except UnicodeError:
-                        print(f'\nSkipping {msg.from_values.full} {msg.subject}', file=sys.stderr)
+                if overwrite:
+                    with open(filepath, 'wb') as file:
+                        try:
+                            email_object = msg.obj.as_string().encode('utf-8')
+                            file.write(email_object)
+                        except UnicodeError:
+                            print(f'\nSkipping {msg.from_values.full} {msg.subject}', file=sys.stderr)
 
-                email_timestamp = msg.date.timestamp()
-                os.utime(filepath, (email_timestamp, email_timestamp))
+                    email_timestamp = msg.date.timestamp()
+                    os.utime(filepath, (email_timestamp, email_timestamp))
 
                 pbar.update(1)
 
     with open(os.path.join(directory, 'senders.txt'), 'w') as file:
         for sender in sorted(senders, key=attrgetter('domain')):
-            domain = sender.email.split('@')[-1]
-            print(f'{sender.email};{domain};{sender.name}', file=file)
+            email = sender.email
+            name = sender.name
+            domain = sender.domain
+            count = senders[sender]
+            print(email, name, domain, count, file=file, sep=";")
 
 
 if __name__ == "__main__":
